feat(c05): migrate durable reservation money to NUMERIC(12,6) (D2)
The money contract's stage 2 / D2: the durable monthly-reservation money
columns were DOUBLE PRECISION while the in-request ledger is integer
micro-USD and the storage floor is NUMERIC(12,6). A month of summed
reservations therefore accumulated binary-float error (M2).

Migration 0053 migrates all six columns across the three tables
(baseline_spend_usd, reserved_usd, settled_spend_usd, estimated_usd,
settled_usd, delta_usd) via ADD NUMERIC -> ROUND backfill -> DROP double ->
RENAME, preserving the non-negative CHECKs and restoring the DEFAULT 0 on
the two scope-month accumulators. delta_usd stays signed (no non-negative
check) as before. budget_reservation.rs decodes through explicit ::float8
casts (the established NUMERIC-money pattern), so Rust arithmetic is
unchanged while the DURABLE value is exact.

A new migrations guard test asserts every one of the six columns is
NUMERIC(12,6), so a future table rebuild cannot silently reintroduce the
float type.

Verified on a fresh disposable PostgreSQL 18.4 + pgvector cluster:
migrations apply + down/up round-trip; tt-core lib 1,132 passed / 0 failed;
the DB-gated budget reservation, key-cap and migrations suites pass. The
DB test caught a real defect during development (the ADD/DROP/RENAME shape
dropped the DEFAULT 0), now fixed.

+++ crates/core/src/budget_reservation.rs
@@ -252,7 +252,7 @@ impl PostgresBudgetReservationStore {
         month_start: NaiveDate,
     ) -> Result<f64, BudgetReservationError> {
         let (baseline, reserved, settled): (f64, f64, f64) = sqlx::query_as(
-            "SELECT baseline_spend_usd, reserved_usd, settled_spend_usd \
+            "SELECT baseline_spend_usd::float8, reserved_usd::float8, settled_spend_usd::float8 \
              FROM gateway_budget_scope_months \
              WHERE scope_kind = $1 AND scope_id = $2 AND month_start = $3 \
              FOR UPDATE",
@@ -362,7 +362,7 @@ impl PostgresBudgetReservationStore {
         }
 
         let stale: Vec<(Uuid, Uuid, Uuid, f64, bool, bool)> = sqlx::query_as(
-            "SELECT id, org_id, api_key_id, estimated_usd, reserves_org, reserves_api_key \
+            "SELECT id, org_id, api_key_id, estimated_usd::float8, reserves_org, reserves_api_key \
              FROM gateway_budget_reservations \
              WHERE month_start = $3 AND status = 'active' AND lease_expires_at <= $4 \
                AND (NOT reserves_org OR $5) \
@@ -611,8 +611,8 @@ impl BudgetReservationStore for PostgresBudgetReservationStore {
 
         let mut tx = self.pool.begin().await?;
         let initial: Option<ReservationRow> = sqlx::query_as(
-            "SELECT org_id, api_key_id, month_start, estimated_usd, \
-                    reserves_org, reserves_api_key, status, settled_usd \
+            "SELECT org_id, api_key_id, month_start, estimated_usd::float8, \
+                    reserves_org, reserves_api_key, status, settled_usd::float8 \
              FROM gateway_budget_reservations WHERE id = $1",
         )
         .bind(reservation.id)
@@ -645,8 +645,8 @@ impl BudgetReservationStore for PostgresBudgetReservationStore {
             status,
             prior_settled,
         ): ReservationRow = sqlx::query_as(
-            "SELECT org_id, api_key_id, month_start, estimated_usd, \
-                    reserves_org, reserves_api_key, status, settled_usd \
+            "SELECT org_id, api_key_id, month_start, estimated_usd::float8, \
+                    reserves_org, reserves_api_key, status, settled_usd::float8 \
              FROM gateway_budget_reservations WHERE id = $1 FOR UPDATE",
         )
         .bind(reservation.id)
